+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 Version 8.4.0-dev
 - allowed setting base url for Nette\Application\LinkGenerator from Nette DIC extension
+- added argument --bootstrap for mytester.php
 
 Version 8.3.0
 - added arguments --coverage and --results for mytester.php